+++ CHANGES.md
@@ -3,6 +3,7 @@ merlin NEXT_VERSION
 
   + merlin binary
     - Add a "heap_mbytes" field to Merlin server responses to report heap usage (#1717)
+    - Add `signature-help` command (#1720)
 
 merlin 4.13
 ===========